% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report-methods.R
\name{addReportFigure}
\alias{addReportFigure}
\title{addReportFigure}
\usage{
addReportFigure(report, description, repPlots, setting.names = list(),
  selected.image = as.integer(1))
}
\arguments{
\item{report}{Report to write the text to.}

\item{description}{Human-readable description of the figure. This must be a non-empty \code{character} vector. The
elements of this vector are concatenated without a separator to form the full description.}

\item{repPlots}{Object of type \code{\linkS4class{ReportPlot}}, or a list of such objects.}

\item{setting.names}{List of plot file element descriptors. Every variable elements in the plot file names must be
included in this list. Set this to empty list if no variable elements are present, that is, if
the figure should present a single report plot.}

\item{selected.image}{Index of plot to be initially selected in the figure.}
}
\value{
The modified report.
}
\description{
Generates HTML code for a figure in the specified report. A figure is a collection of images (plots), of which only
one is visible at any given moment.
}
\seealso{
\code{\link{addReportTables}} for adding a listing of tables; \code{\linkS4class{Report}} for other functions
  adding contents to an HTML report
}
\author{
adapted by Fabian Mueller from RnBeads code by Yassen Assenov
}
