% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportPlot-methods.R
\name{getReportPlotFiles}
\alias{getReportPlotFiles}
\title{getReportPlotFiles}
\usage{
getReportPlotFiles(repPlot)
}
\arguments{
\item{repPlot}{Report plot of interest. This must be an object of type \code{\linkS4class{ReportPlot}}.}
}
\value{
Non-empty \code{character} vector of absolute file names.
}
\description{
Gets the list of all files that are planned to be generated, or were already generated by the given report plot.
}
\examples{
\donttest{
plot.image <- createReportPlot('scatterplot', high.png = 200)
getReportPlotFiles(plot.image)
}
}
\author{
Yassen Assenov
}
