% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report-methods.R
\name{getReportReference}
\alias{getReportReference}
\title{getReportReference}
\usage{
getReportReference(report, txt)
}
\arguments{
\item{report}{Report that contains the reference to be cited.}

\item{txt}{Text of the reference in the form of a non-empty \code{character} vector. This reference must already
added to the report.}
}
\value{
Citation of the reference item (including a link) in the form of a one-element \code{character} vector. If
        the specified reference item is not found in the report, this method returns an empty string.
}
\description{
Creates a string that points to the given reference item in the specified report.
}
\examples{
\donttest{
report <- createReport("example.html", "Example", init.configuration = TRUE)
txt.reference <- c("Bird A. ", "<i>Nucleic Acids Res.</i> <b>8</b> (1980)")
report <- addReportReference(report, txt.reference)
txt <- c("This was shown in ", getReportReference(report, txt.reference), ".")
addReportParagraph(report, txt)
}
}
\seealso{
\code{\link{addReportReference}} for adding a reference item to a report; \code{\linkS4class{Report}} for
  other functions adding contents to an HTML report
}
\author{
adapted by Fabian Mueller from RnBeads code by Yassen Assenov
}
