% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggMsgPlot}
\alias{ggMsgPlot}
\title{ggMsgPlot}
\usage{
ggMsgPlot(txt)
}
\arguments{
\item{txt}{Text to be plotted.}
}
\value{
The newly initialized \code{ggplot} instance.
}
\description{
Creates a plot, using \pkg{ggplot2}, with a single text message.
}
\examples{
\donttest{
x11(width = 5, height = 5)
ggMsgPlot("Missing data")
}
}
\author{
adapted by Fabian Mueller from RnBeads code by Yassen Assenov
}
