% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report-methods.R
\name{initializeReports}
\alias{initializeReports}
\title{initializeReports}
\usage{
initializeReports(reportDir, configDir = "_config",
  logo = system.file(file.path("extdata", "logo.png"), package = "muReportR",
  mustWork = TRUE))
}
\arguments{
\item{reportDir}{Directory to host report files. This must be a \code{character} of length one that specifies
a non-existent path, as this methods attempts to create it.}

\item{configDir}{Subdirectory to host configuration files shared by the reports. This must be a
\code{character} of length one that gives location as a path relative to \code{reportDir}.
Also, strong restrictions apply to the path name. See the description of the
\code{\link{createReport}} function for more details. This method creates the directory and
copies configuration files that define cascading style sheet (CSS) definitions and
Javascript functions used by the HTML reports.}
}
\value{
\code{TRUE} if the report directory was successfully created and the configuration files were copied to the
        specified location; \code{FALSE} otherwise.
}
\description{
Creates a new directory to host HTML reports and copies the shared configuration files.
}
\examples{
\donttest{
reportDir <- "~/infinium_studies/cancer_study/reports"
if (!initializeReports(reportDir)) {
    cat("ERROR: Could not initialize configuration in ", reportDir, "\\n", sep = "")
}
}

}
\seealso{
\code{\link{createReport}} for initializing an HTML report
}
\author{
Yassen Assenov
}
