% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muReportR.R
\docType{package}
\name{muReportR}
\alias{muReportR}
\alias{muReportR-package}
\title{muReportR: Generating analysis reports in R.}
\description{
Generate HTML reports from R. Based on code from the RnBeads package.
}
